Commit before my holiday :3

+++ src/compiler/types.ts
@@ -74,7 +74,7 @@ export enum CatnipValueFlags {
 export enum CatnipValueFormat {
     i32,
     HSTRING_PTR, // (i32)
-    VALUE_BOXED, // (i64)
+    VALUE_BOXED, // (f64)
     f64,
     ANY
 }
@@ -87,7 +87,7 @@ export function getValueFormatSpiderType(inputFormat: CatnipValueFormat): Spider
         case CatnipValueFormat.f64:
             return SpiderNumberType.f64;
         case CatnipValueFormat.VALUE_BOXED:
-            return SpiderNumberType.i64;
+            return SpiderNumberType.f64;
         case CatnipValueFormat.ANY:
             throw new Error("Cannot convert format 'any' to a number type.");
     }
